@@ -5,8 +5,8 @@
   "dependencies": {
     "express": "^4.21.2",
     "process": "^0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "webpack": "^5.97.1"
